rename ts loop to time_series; add cnn decoder
