+++ lightwood/encoders/time_series/__init__.py
@@ -1,2 +1,2 @@
-from lightwood.encoders.time_series.rnn import TimeSeriesEncoder
+from lightwood.encoders.time_series.time_series import TimeSeriesEncoder
 from lightwood.encoders.time_series.plain import TimeSeriesPlainEncoder
